add withdraw & extendownership actions + merge registerownership and extendownership modals
