+++ src/scss/_custom.scss
@@ -64,6 +64,10 @@ body {
     border: none;
   }
 
+  .listing-img-card svg {
+    width: 100%;
+  }
+
   .listing-img-card img {
     border-radius: 20px 20px 0 0;
   }
