--- src/shared/models/listingInfo.model.ts
@@ -1,11 +1,0 @@
-export interface IListingInfo {
-  value: string;
-  rewardRate: string;
-  currentOwner: string;
-  rewardPool: string;
-  rewardPoolOfList: string;
-  totalToken: string;
-  ownerWalletId: string;
-  ownTimeLeft?: string;
-  tokenRecharged?: string;
-}
